+++ scripts/docker_image_entry.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-echo "WARNING: This Docker image will be removed permanently on 2024/12/1. Please STOP using this image ASAP"
-exec java -jar /dxCompiler.jar "$@"